More help work, improved jog control and Go Live button
git-svn-id: https://sbgracing.sourcerepo.com/sbgracing/RacePad/RacePad@137 9925204c-63e1-4f11-ae18-372a25b3a23a

+++ Classes/UISupport/RacePadViewController.h
@@ -29,13 +29,13 @@
 	CGPoint downPoint;
 	bool initialised;
 	float lastAngle;
-	float direction;
+	int direction;
 }
 
 @property (readonly) CGPoint downPoint;
 @property (nonatomic) bool initialised;
 @property (nonatomic) float lastAngle;
-@property (nonatomic) float direction;
+@property (nonatomic) int direction;
 
 - (float)angleOfPoint:(CGPoint)point InView:(UIView *)gestureView;
